Add account menu to sidebar on settings page
