+++ shared/i18n/locales/en_US/translation.json
@@ -149,11 +149,15 @@
   "Toggle sidebar": "Toggle sidebar",
   "Delete {{ documentName }}": "Delete {{ documentName }}",
   "Return to App": "Back to App",
+  "Installation": "Installation",
+  "No results": "No results",
+  "Previous page": "Previous page",
+  "Next page": "Next page",
   "Account": "Account",
   "Notifications": "Notifications",
   "API Tokens": "API Tokens",
-  "Team": "Team",
   "Details": "Details",
+  "Team": "Team",
   "Security": "Security",
   "Features": "Features",
   "Members": "Members",
@@ -162,10 +166,6 @@
   "Import": "Import",
   "Export": "Export",
   "Integrations": "Integrations",
-  "Installation": "Installation",
-  "No results": "No results",
-  "Previous page": "Previous page",
-  "Next page": "Next page",
   "Insert column after": "Insert column after",
   "Insert column before": "Insert column before",
   "Insert row after": "Insert row after",
